+++ Missile Champions/Missile Champions.vcxproj
@@ -122,6 +122,7 @@
     <ClCompile Include="Assets.cpp" />
     <ClCompile Include="Graphics.cpp" />
     <ClCompile Include="MChamps.cpp" />
+    <ClCompile Include="Objects\Ball.cpp" />
     <ClCompile Include="Objects\BoostStreak.cpp" />
     <ClCompile Include="Objects\Car.cpp" />
     <ClCompile Include="OnCleanup.cpp" />
