+++ Missile Champions/Missile Champions.vcxproj.filters
@@ -74,6 +74,9 @@
     <ClCompile Include="Objects\BoostStreak.cpp">
       <Filter>Source Files</Filter>
     </ClCompile>
+    <ClCompile Include="Objects\Ball.cpp">
+      <Filter>Source Files</Filter>
+    </ClCompile>
   </ItemGroup>
   <ItemGroup>
     <Image Include="res\img\title_screen.png">
